+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2025-09-16
+
+### Added
+- Support for initializing MessageClient using `CLOUDMAILIN_SMTP_URL` environment variable
+- Support for passing `smtpUrl` directly as an option in MessageClientOptions
+
+### Changed
+- MessageClient constructor signature simplified to accept `MessageClientOptions` (where all fields are now optional)
 
 ## [0.1.0] - 2024-01-23